requirements.txt
asgiref==3.4.1
black==21.9b0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Django==3.2.8
django-crontab==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
idna==3.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.1
pytz==2021.3
regex==2021.10.8
requests==2.26.0
sqlparse==0.4.2
tomli==1.2.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
